backend/requirements.txt
alembic==1.17.1
asyncpg==0.30.0
fastapi==0.121.3
psycopg2-binary==2.9.11
pydantic==2.12.4
python-dotenv==1.2.1
SQLAlchemy==2.0.44
uvicorn[standard]==0.38.0
websockets==14.1
supabase==2.10.0
python-multipart==0.0.9